--- src/Foundation/Config.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Laraver\Waimai\Foundation;
-
-use Illuminate\Support\Collection;
-
-class Config extends Collection
-{
-}
